my_application_python/24.1_course_drf/courses/serializers.py
from rest_framework import serializers

from courses.models import Course, Lesson


class CourseSerialaizer(serializers.ModelSerializer):

    class Meta:
        model = Course
        fields = '__all__'


class LessonSerialaizer(serializers.ModelSerializer):

    class Meta:
        model = Lesson
        fields = '__all__'
